status: move into lightningd/status.
It's really a lightningd-only thing, and we're about to do surgery on it.
