+++ lightningd/cryptomsg.c
@@ -7,8 +7,8 @@
 #include <ccan/short_types/short_types.h>
 #include <ccan/take/take.h>
 #include <lightningd/cryptomsg.h>
+#include <lightningd/status.h>
 #include <sodium/crypto_aead_chacha20poly1305.h>
-#include <status.h>
 #include <utils.h>
 #include <wire/wire.h>
 #include <wire/wire_io.h>
