+++ lightningd/status.c
